port buildSrc to build-logic
