+++ gradle/libs.versions.toml
@@ -36,6 +36,18 @@ hid4java = { module = "org.hid4java:hid4java", version.ref = "hid4java" }
 # Submodule: DesktopDemo
 jsr305 = { module = "com.google.code.findbugs:jsr305", version.ref = "jsr305" }
 
+# Build plugins (for build-logic)
+spotbugs-gradle-plugin = { module = "com.github.spotbugs.snom:spotbugs-gradle-plugin", version.ref = "spotbugs-plugin" }
+android-gradle-plugin = { module = "com.android.tools.build:gradle", version.ref = "android-gradle" }
+spotless-plugin-gradle = { module = "com.diffplug.spotless:spotless-plugin-gradle", version.ref = "spotless" }
+kotlin-gradle-plugin = { module = "org.jetbrains.kotlin:kotlin-gradle-plugin", version.ref = "kotlin" }
+
+# Build dependencies
+spotbugs = { module = "com.github.spotbugs:spotbugs", version.ref = "spotbugs-core" }
+spotbugs-annotations = { module = "com.github.spotbugs:spotbugs-annotations", version.ref = "spotbugs-core" }
+findsecbugs-plugin = { module = "com.h3xstream.findsecbugs:findsecbugs-plugin", version.ref = "findsecbugs" }
+slf4j-api = { module = "org.slf4j:slf4j-api", version.ref = "slf4j" }
+
 # Shared
 androidx-multidex = { module = "androidx.multidex:multidex", version.ref = "multidex" }
 androidx-junit = { module = "androidx.test.ext:junit", version.ref = "androidx-junit" }
@@ -50,8 +62,16 @@ mockito = { module = "org.mockito:mockito-core", version.ref = "mockito" }
 
 [plugins]
 kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
+android-application = { id = "com.android.application", version.ref = "android-gradle" }
+android-library = { id = "com.android.library", version.ref = "android-gradle" }
 
 [versions]
+android-gradle = "8.13.1"
+spotbugs-plugin = "6.2.2"
+spotbugs-core = "4.9.8"
+findsecbugs = "1.14.0"
+spotless = "8.0.0"
+slf4j = "2.0.17"
 annotations = "26.0.2-1"
 appcompat = "1.7.1"
 bcpkix-jdk15to18 = "1.82"
@@ -75,7 +95,7 @@ logback-classic = "1.5.20"
 material = "1.13.0"
 moshi = "1.15.2"
 multidex = "2.0.1"
-navigation-fragment-ktx = "2.9.5"
+navigation-fragment-ktx = "2.9.6"
 recyclerview = "1.4.0"
 androidx-annotation = "1.9.1"
 androidx-junit = "1.3.0"
