--- pom.gradle.kts
@@ -1,43 +1,0 @@
-/*
- * Copyright (C) 2025 Yubico.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import org.gradle.api.publish.maven.MavenPom
-import org.gradle.api.Action
-
-extra["pomData"] = Action<MavenPom> {
-    name.set(extra["pomName"].toString())
-    description.set(project.description)
-    url.set("https://github.com/Yubico/yubikit-android/tree/main/${project.name}")
-    licenses {
-        license {
-            name.set("The Apache License, Version 2.0")
-            url.set("http://www.apache.org/licenses/LICENSE-2.0.txt")
-        }
-    }
-    developers {
-        developer {
-            id.set("dainnilsson")
-            name.set("Dain Nilsson")
-            email.set("dain@yubico.com")
-        }
-    }
-    scm {
-        connection.set("scm:git:https://github.com/Yubico/yubikit-android.git")
-        developerConnection.set("scm:git:ssh://github.com/Yubico/yubikit-android.git")
-        url.set("https://github.com/Yubico/yubikit-android")
-    }
-}
-
