+++ settings.gradle.kts
@@ -15,14 +15,11 @@
  */
 
 pluginManagement {
+    includeBuild("build-logic")
     repositories {
         mavenCentral()
         google()
         gradlePluginPortal()
-    }
-    plugins {
-        id("com.android.application") version "8.13.0"
-        id("com.android.library") version "8.13.0"
     }
 }
